requirements.txt
h5py==2.9.0
tqdm
tensorflow==1.13.0rc2
dm-sonnet==1.27
numpy==1.15.4
matplotlib
jupyter
pyyaml
wrapt==1.11.2
tensorflow-probability==0.5.0
librosa
